fix(storage): sstables merging logic
Every time we reached to sstable count limit, the program would always
pick the first two and merge them. But that leaves us with a problem:
recently created sstables are always put at the end as they are the most
recent ones, so essentially we were only merging the biggest sstables at
the start and leaving the rest of the space to the smaller ones. Now, every
time a new sstable is about to be created, if we are already at the
limit, we merge the new sstable with the oldest one resulting in a much
more homogeneous distribution of the data